+++ ajax/logviewer.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * ---------------------------------------------------------------------
  *
@@ -32,6 +31,8 @@
  *
  * ---------------------------------------------------------------------
  */
+
+use Glpi\System\Log\LogViewer;
 
 include('../inc/includes.php');
