+++ front/logs.php
@@ -33,6 +33,8 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\System\Log\LogViewer;
+
 include('../inc/includes.php');
 
 Session::checkRight("logs", READ);
