+++ src/System/Log/LogViewer.php
@@ -33,8 +33,13 @@
  * ---------------------------------------------------------------------
  */
 
+namespace Glpi\System\Log;
+
+use CommonGLPI;
 use Glpi\Application\View\TemplateRenderer;
 use Glpi\Event;
+use Session;
+use Toolbox;
 
 final class LogViewer extends CommonGLPI
 {
