+++ tests/units/Html.php
@@ -338,7 +338,7 @@ class Html extends \GLPITestCase
             'Rule',
             'Profile',
             'QueuedNotification',
-            'LogViewer',
+            'Glpi\System\Log\LogViewer',
             'Glpi\Inventory\Inventory'
         ];
         $this->string($menu['admin']['title'])->isIdenticalTo('Administration');
